Перенёс модели в отдельные файлы. Осталось связать классы.
