+++ MovieQuiz/Models/QuizResultsViewModel.swift
@@ -1,0 +1,2 @@
+
+import Foundation
